Let ffmpeg directly read files if possible
Instead of always using pipes to feed ffmpeg the input file,
if we have a file on disk, just let ffmpeg open and read it by itself.

ffmpeg needs to be able to seek on certain formats to read them, so
we can't use pipes. This commit fixes reading those formats.