Fallback to using the wave module if scipy raises some exceptions
scipy only supports wav files with a bit depth in (8, 16, 32, 64, 96, 128)
so in other cases (most notably, 24), it raises a ValueError("Unsupported "
" bit depth: the wav file has {}-bit data.".format(bit_depth)) exception.

Also, When the data doesn't start with the standard RIFF header, scipy
raises a ValueError("File format {}... not understood.".format(repr(data)))
exception.

In those cases, as well as when scipy is not available, fallback to
using the python standard wave module, just in case it does support
reading that data (for some cases of 24 bit depth data, it does).