Add mediainfo_json function that returns information about a media file
The current mediainfo function mixes information from different streams so
mediainfo_json is a replacement which just asks avprobe/ffprobe to
return the information as a json dictionary and parses it, thus giving
separate information for all streams

Additionally, avprobe doesn't return as much information as ffprobe on
the json data, so we extract that missing information from its stderr
output.

Finally, also, remove an unneeded import from the code of mediainfo.