Add support to read audio using 24/32 bits per sample
By default, avconv/ffmpeg generates pcm_s16le data when converting
to a wav file, even if the input is using 24/32 bits. For this
reason, we have to find out the bits_per_sample of the input data
and then request the same value in the output file.

Note that the wave module doesn't support a frequency over 44 KHz
and scipy doesn't support 24 bits, so we convert to 32 bits in that
case (sample_fmt would be 's32' while bits_per_sample would be 24)
to not lose any information.