Add fsdecode as a wrapper for os.fsdecode and clean filename handling
os.fsdecode was introduced in python 3.2, so I implemented fsdecode in
utils.py to be able to use it on previous versions. fsdecode thus
returns the filename passed as parameter as a unicode string (either
if it's str or bytes), and if a PathLike object is used, it returns
its filename. In other case, it raises a TypeError exception.

Changed code in from_file and mediainfo_json to use it, which simplifies
the code a bit and hopefully fixes the last windows appveyor error.