Use the cache protocol and read_ahead_limit when piping data to ffmpeg
When piping the input file to ffmpeg on stdin ffmpeg can't seek on the
file so it's unable to read the file in advance and has to guess some
parameters. This is specially worrying on mp3 files, where it's common
that reading the input from a file and piping the same file on stdin
gives different wave results. To fix this, we now use the cache protocol
on stdin, so ffmpeg caches the input and also set read_ahead_limit to -1
(by default), so the cache can hold the whole file.

This might make ffmpeg use more memory (since it has to hold the whole
file in memory), so read_ahead_limit can be set to something else when
calling `from_file``` although of course, then the resulting wave data
is not guaranteed to be 100% accurate.