Do not close from_file input files if we didn't open them
from_file* methods allow to be passed a path to a file to read, but also
a filelike object like a BytesIO object with the file contents. In that
case, we shouldn't close the object, since that makes it unavailable
for the caller if it needs to be used for something else after we finish
processing it.

This commit changes _fd_or_path_or_tempfile so it doesn't only return
a file descriptor but also a boolean that tells the caller if the file
descriptor can be closed or not.