compare dBFS against dBFS from amplitude
Since AudioSegment.dBFS calculates the value based on the amplitude,
we should compare against that in our tests.
The expected value for the 8-bit test had to be changed because audioop.rms truncates its result to an integer,
and so the dBFS does not precisely match the value for the 16-bit version.