Fix imports of AudioSegments in stereo_to_ms and ms_to_stereo
There is no symbol AudioSegment defined in these functions. Calling them results in:
  NameError: name 'AudioSegment' is not defined. Did you mean: 'audio_segment'?
We cannot import AudioSegment at the top of the module, so we must import it inside of this function.
To be consistent with the rest of the function, I'm using tabs as indentation here. We could consider turning the indentation into spaces.